feat: create a HeartIcon for reusability and easier implementation for wishlist.
